split edit portfolio into components, add ksbs
