+++ app/_components/Input.tsx
@@ -1,0 +1,20 @@
+export function Input({
+  value,
+  onChange,
+  onBlur,
+  className,
+}: {
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+  onBlur: () => void;
+  className?: string;
+}) {
+  return (
+    <input
+      className={"p-2 w-[90%] rounded " + className}
+      value={value}
+      onChange={onChange}
+      onBlur={onBlur}
+    />
+  );
+}
